Fix init-aws.js to behave as expected
There were some patterns here that were mimicking what the AWS SDK already does for you automatically, but it was only mimicking a subset of the behavior. It seems the only custom thing that is needed is to use `profile` and `region` settings in the arc config file, which I kept.

Update changelog.md

rm trailing space

working test suite