Support AWS SSO profiles
We should not require or encourage users to keep long-lived credentials
with high privileges on local disk.

I use AWS SSO (now called AWS IAM Identity Center).

This is the minimal change I had to make to make it possible to use
AWS SSO profiles with Architect.

It detects the presence of the sso_start_url key in the profile and
uses the SsoCredentials provider.  This provider was introduced in
aws-sdk 2.1093.0 [1].

However this change is incomplete on its own:

- for this to work, the top-level package should require at least aws-sdk 2.1093.0
  (currently it requires 2.1001.0).  I can raise this PR if it is likely
  to be accepted.
- the documentation should be updated to make it clear how to use this.
  likewise, I can raise this PR if it is likely to be accepted.

There are also some remaining niggles with how this works:

- it requires the profile to be in ~/.aws/credentials.  I don't have this
  file at all, because I don't have any long-lived credentials.  I use
  ~/.aws/config alone.  Really, for SSO credentials (and also
  credential_process credentials) it should be acceptable to store the
  profile entirely in ~/.aws/config.
- it requires the profile header to be literally `[profile-name]`.  If the
  header is `[profile profile-name]`, the profile is not detected by init-aws.js
  even though it is detected by aws-sdk itself.

I don't really understand what all this detection is trying to do.  It looks
like it's trying to second-guess the AWS SDK, but doing a bad job of it and so
failing to actually support half of what the AWS SDK does.

[1]: https://github.com/aws/aws-sdk-js/blob/master/CHANGELOG.md#210930